app-text/katarakt: Katarakt is a simple PDF viewer (new package)
katarakt is a simple PDF viewer. It is designed to use as much available screen
space as possible.

There are currently two layouts. The presentation layout is very simple and
only supports scrolling on a per page basis. As the name suggests the current
page is displayed in the center and zoomed to fit the window. It is active by
default.

The grid layout is much more advanced and offers continuous (smooth, per pixel)
scrolling, zooming and adjusting the column count. Pages keep their correct
relative size and are shown in a grid.

Package-Manager: Portage-2.3.84, Repoman-2.3.20
Signed-off-by: Kurt Kanzenbach <kurt@kmk-computers.de>