dev-util/quilttools: Mailbox to quilt series converter (new package)
mb2q analyses the complete mailbox and collects various tags (Reviewed-by,
Acked-by, Tested-by) from replies to individual patches and to the cover
letter. If a tag is in a reply to the cover letter it is applied to all
patches which reference the cover letter.

Optionally the output can be redirected into a mailbox which can be applied
with git-am.

Package-Manager: Portage-2.3.84, Repoman-2.3.20
Signed-off-by: Kurt Kanzenbach <kurt@kmk-computers.de>