dev-python/aiofiles: enable tests
had to switch to the github tarball because the pypi tarball
does not contain tests. Also pytest-asyncio sadly does not
support py3_8 or pypy3 yet, so that had to go.

If it is desirable to add py3_8 or pypy3 anyway.
A workaround could be to RESTRICT="python_targets_python3_8? ( test )"
and REQUIRED_USE="python_targets_python3_8? ( !test )"

Package-Manager: Portage-2.3.93, Repoman-2.3.20
Signed-off-by: Andrew Ammerlaan <andrewammerlaan@riseup.net>