app-editors/bluegriffon-bin: HTML and EPUB editor
New ebuild, I'm not sure about proper category
May be it should be www-apps, however, it can be used
as non-web editor (epub editor)
Binary package because building from source is terrible hell

From Wikipedia:
BlueGriffon is a WYSIWYG content editor for the World Wide Web.
It is based on the discontinued Nvu editor,
which in turn is based on the Composer
component of the Mozilla Application Suite. Powered by Gecko,
the rendering engine of Firefox, it can edit Web pages
in conformance to Web Standards.
It runs on Microsoft Windows, macOS and Linux.,

Package-Manager: Portage-2.3.89, Repoman-2.3.20
Signed-off-by: Denis Reva <denis7774@gmail.com>