guru/profiles/package.mask: after full test kikoplay works
We can't simply and roughly mask a software that depends on
slotted lua without detailed investigation, because even in
::gentoo there are still many softwares that depend on
lua:5.1, lua:5.2, lua:5.3, now lua5.4 is released.
This is the discussion:
https://github.com/gentoo/guru/issues/9
This is extremely strange and unreasonable for blocking
lua:5.1, lua:5.2, lua:5.3 in ::gentoo,
Many excellent software which using latest lua may blocked.

Package-Manager: Portage-2.3.99, Repoman-2.3.22
Signed-off-by: Huang Rui <vowstar@gmail.com>