.github/workflows/repoman.yml: Use script to auto run the latest rm
By using a seperate script instead of running everything in the yml
file, we gain the ability to use and manipulate variables. This
means that we can now automatically run the latest version of
repoman, which should make this more hands-off by removing the
need to update this every now and then.

Signed-off-by: Andrew Ammerlaan <andrewammerlaan@riseup.net>