media-libs/libbpg: Add missing dependencies.
libnuma support is optional. If sys-process/numactl is installed, it
will be used, if not, not. The check for libnuma happens in CMake, which
is called by the Makefile. But the Makefile also needs to know if it
should link in libnuma or not. No one tells it. I chose the easy way out
of this pickle, made libnuma mandatory and added `-lnuma` to LIBS.

Closes: https://bugs.gentoo.org/781113
Signed-off-by: Ronny (tastytea) Gutbrod <gentoo@tastytea.de>