net-im/abaddon: dont use git-r3
Not live packages (non *9999) shouldn't inherit git-r3. We can use
GitHub's snapshots to receive a tarball of simpleini at the wanted
commit, and then link it in the wanted position by abaddon.

Note: here I'm using bundled software, which is discouraged in
Gentoo. simpleini should become a new package, on which abaddon
depends.

Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>