dev-R/lattice: fix and improve src_uri, add optfeature for dev-R/MASS
* Improve SRC_URI to use ver_rs and not a hard coded URI
* Add optfeature for dev-R/MASS

Currently, dev-R/lattice will not fetch because the source for
version 0.20.44 has been archived. Therefore, the package must use
the URL for archived versions. In addition, the ver_rs function is used
because the URI for the package replaced the second period with a '-'
in the version number. Since https://cran.r-project.org/ is listed in
thirdpartymirrors, it should be referenced in SRC_URI with a mirror:///
URI. In addition, dev-R/MASS will enable optional features so an
optfeature function is added. Also, the HOMEPAGE variable was set.

Closes: https://bugs.gentoo.org/815781
Package-Manager: Portage-3.0.20, Repoman-3.0.3
Signed-off-by: Lucas Mitrak <lucas@lucasmitrak.com>