gui-apps/foot: Support ncurses terminfo
The gui-apps/foot-terminfo package conflicts with versions of ncurses
past 2021-07-31. To prevent conflicts, this package will not install the
separate terminfo if a recent version of ncurses is installed.

This requires users to manually remove gui-apps/foot-terminfo before
upgrading sys-libs/ncurses. I would prefer if that wasn't necessary, but
I'm not sure how to get around it.

Package-Manager: Portage-3.0.28, Repoman-3.0.3
Signed-off-by: Ryan Fox <flewkey@2a03.party>