gui-apps/foot-terminfo: Prevent package collision
Versions of sys-libs/ncurses since 2021-07-31 have added a version of
the foot terminfo with less features. Due to issues with installing
terminfo in a nonstandard location, the current recommendation is to
use the ncurses version if it is installed.

This adds a blocking dependency for recent versions of sys-libs/ncurses
to prevent the package conflict.

Closes: https://bugs.gentoo.org/825434
Package-Manager: Portage-3.0.28, Repoman-3.0.3
Signed-off-by: Ryan Fox <flewkey@2a03.party>