app-editors/mined: Add new package
This adds a package for the MinEd text editor.
The test suite is RESTRICT-ed because it will throw a QA message if
used, due to an upstream make jobserver issue. If removed, it will
still complete, but a QA error will show.
The make process uses the -j1 option for the same reason.

Package-Manager: Portage-3.0.28, Repoman-3.0.3
Signed-off-by: Alex Holcomb <alex@alexsdigital.rodeo>