sci-physics/SU2: revert unbundling metis/parmetis (as fail tests)
The unbundling of metis/[armetis results in numerous failures
of tests, especially the Tutorial tests that are preferred
to be checked.

The revision with additional USE-flags is now '-r1xx'.

The changes also enable python tests;
add restriction '<dev-cpp/catch-3:0' with catch to be required
only with USE="test";

The USE="openmp" results in inf. run of Tutorial and TestCases tests
and USE="mixed-precision" results in test failures.
Assumed that this USE-flags should be maske, especcialy USE="openmp".

Signed-off-by: Sergey Torokhov <torokhov-s-a@yandex.ru>