games-emulation/dosbox-x: Update for upstream's rolling release adoption
Without a clear explanation, the upstream of this package has stopped
tagging releases and willy-nilly switched to "continuously built"
"nightly builds" (a.k.a. prone-to-break-on-any-day binaries put for user
downloads at an obscure place).  To allow Gentoo users to use recent
versions of this package without having to resort to the live ebuild, we
have no choice other than to make new non-live ebuilds based on GitHub
snapshot tarballs from now on.

Signed-off-by: Yuan Liao <liaoyuan@gmail.com>