app-editors/pulsar-bin: Support arm64
Tested ebuild on an arm64 prefix. Couldn't do a proper run test
because that machine is awfully slow.

Also removed the redundant chmods, if we could simply mv the files
into image instead of using doins, the modes from the original
tarball are kept and we don't need to chmod.

Signed-off-by: YiFei Zhu <zhuyifei1999@gmail.com>