net-im/mautrix-discord: add lottieconverter rdep
These versions of mautrix discord will crash on runtime if
lottieconverter is not in PATH when it tries bridging certain
stickers, which is why I made it an rdep. The crashing is
currently fixed in git master, which for the next release
lottieconverter would be more suitable for an optfeature
I suppose.

Signed-off-by: Richard-Rogalski <rrogalski@tutanota.com>