dev-ml/ollama: move log ownership changes to pkg_preinst
Since the ollama UID and GID are dynamically assigned, we cannot rely on
them being the same on different machines. Calling fowners ollama:ollama
<file> in src_install could cause permission issues at runtime, when
e.g. a binary package is created on one machine and used on another.

Moving the offending code to pkg_preinst, which is called for binary and
source packages, solves the issue.

Closes: https://github.com/gentoo/guru/pull/210
Signed-off-by: Lucio Sauer <watermanpaint@posteo.net>