net-im/mautrix-whatsapp: 0.10.9 version bump.
Old versions not working anymore, so removing them.

Also changed log path from /var/log/mautrix to /var/log/mautrix/whatsapp for
consistency with net-im/mautrix-signal and also because that was always the plan
and a /var/log/mautrix/whatsapp directory was being created for that purpose but
never used because of mistakes in the ebuild.

Signed-off-by: Jesús P Rey (Chuso) <git@chuso.net>