media-gfx/airsane: new package, add 0.4.3
AirSane is SANE frontend that publishes SANE scanners to MacOS,
Android, and Windows via Apple AirScan/eSCL. Do not confuse this
package with media-gfx/sane-airscan or the eSCL backend - those
are eSCL clients. This package creates an eSCL server to
export local legacy scanners to the network via the AirScan/eSCL
protocol.

Signed-off-by: Christopher Byrne <salah.coronya@gmail.com>