games-util/ProtonUp-Qt: fix pyside6 deps on v2.9.2
Launching ProtonUp-Qt with the currently available pyside6 version
(which is 6.8.1 at the time of writing) fails due to missing
Python module imports.

Analysis of the ProtonUp-Qt sourcecode shows that the following
imports are required:
- Pyside6.QtCore
- PySide6.QtGui
- PySide6.QtUiTools
- PySide6.QtWidgets

QtCore is always available in pysid6. Translating them to useflags
they become:
- gui
- uitools
- widgets

Fix the pyside6 dependency accordingly.

Signed-off-by: Tobias Jakobi <tjakobi@math.uni-bielefeld.de>
Signed-off-by: David Roman <davidroman96@gmail.com>