games-util/ProtonUp-Qt: fix dependencies of v2.11.1
Use the version information from the project's requirement.txt
This in particular means using an up-to-date version of both
the steam and the vdf library.

Without these latest libs the application is not able to
access any of the .vdf files of current stable Steam clients
any more.

While at it, at the missing gh part of the download suffix and
break the overly long line.

Signed-off-by: Tobias Jakobi <tjakobi@math.uni-bielefeld.de>