dev-vcs/jj: add USE flags to live ebuild
This adds a git2 USE flag (enabled by default) to
match the git2 feature flag added by upstream.

This also adds a gix-max-performance USE flag to
match an existing featre flag. Upstream recommends
this be enabled by packagers, but I've left it
disabled by default here, as it comes with both a
dependency on cmake (for an unfortunately-vendored
zlib-ng) and potential incompatibilties (since it
uses assembly versions of SHA1 code in gitoxide).

Signed-off-by: demize <demize@unstable.systems>