dev-embedded/esp-idf: update Manifest for esp-idf-5.4.1
Verification failed due to a filesize mismatch.
This zip contains the .git directory of the project. The differences in
size was due to some differences in the git packs. However, we remove
the .git directory, so it doesn't makes any difference. Also, the md5sum
hash of all other files were the same.

Closes: https://bugs.gentoo.org/953048
Signed-off-by: David Roman <davidroman96@gmail.com>