dev-python/pywebview: fix totally broken dependencies on PYTHON_USEDEP
The maintainer of this package appears to have failed to utilize the
basics of python dependency management on gentoo, resulting in
dependencies built for one version of python being imported -- or not --
by this package built for a different version of python.

Upon making this fix, we observe fatal pkgcheck errors. Hold on for a
followup commit.

Fixes: e2d1acff6d607b433c90dc988795730d9593c00f
Fixes: 67365ce1560d06e67e76d0d1ae5e1434249dd1a9
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>