app-emulation/arnold: update SRC_URI, WX_GTK_VER
Tested with https://github.com/llopis/amstrad-diagnostics.
All tests I can run pass--I don't have all the keys needed
for the keyboard test or a joystick, but it seems to work.

The source file is the same, but the old SRC_URI is dead,
thus the Manifest wasn't updated--the hashes match :)

Signed-off-by: Vivian Heisz (demize) <demize@unstable.systems>