games-emulation/rpcs3: Update hidapi unbundling regex
One of the hidapi unbundling steps in the ebuild is to replace
"hidapi.h" with <hidapi/hidapi.h>, but that include line was recently
changed from "hidapi.h" to <hidapi.h> so the regex does nothing because
it specifically looks for and replaces the quotes. This leads to a
compile error claiming hidapi.h could not be found.

The change was made in this commit:
https://github.com/RPCS3/rpcs3/commit/38289f3bc1d711aa8b9d801668f8c9b8059d35df

I am just updating the regex to look for < and > instead.

Signed-off-by: Erica Nebula <EricaNebula@Proton.me>