games-emulation/rpcs3: Unbundle zstd, stb, and openal
These three patches were present on the 0.0.34 ebuild but needed minor changes
to be applied to 0.0.37. I've also applied them to the live ebuild this
time, as they were not before.

The patches for both ebuilds are currently identical but kept in
distinct versioned files for easier handling in case they diverge in the future.

Signed-off-by: Erica Nebula <EricaNebula@Proton.me>