sys-boot/zfsbootmenu: add 3.0.1
Bump the ebuild to the newest version.

Point to the Gentoo wiki instead of the upstream GitHub repository for
configuration instructions. The upstream README doesn't describe the
process at all. The upstream documentation (linked in that README) has
distribution-specific instructions, where Gentoo is not mentioned.

The instruction in Gentoo wiki is complete and based on the information
from the upstream documentation, adjusted to our needs.

Signed-off-by: Michal Rostecki <vadorovsky@disroot.org>