media-sound/spct: add 1.1.7
This is a bit of an overhaul to the ebuild, done because upstream added
a shared/static library version of the application.

With the changes here, prefix builds, multilib builds, and cross builds
should all work correctly (and I've tested to make sure they do).

Support is included for the binary itself, with the shared library gated
behind a USE flag.

Signed-off-by: Vivian Heisz (vhz) <demize@unstable.systems>