Update ollama-0.9.6.ebuild removed link to personal repo.
Why is there a link to your personal repo negril? Ollama provides old versions of their software. They are both even hosted on Github so it's not for backup reasons. Once they stop hosting it your backup repo might be nice. But add it once they stop doing that. As is it looks malicious.