dev-cpp/cpprestsdk: patch for llvm up to 20
Llvm has previously required sizes/implementation of objects used in
containers and std::unique_ptr for more methods than gcc. Also in later
llvms like 19 and 20 unsigned types are not possible to use with the
std::char_traits. Cpprestsdk does use a lot of unsigned types with C++
streams and while that seems to be fine with MSVC and at least gcc up
until some point, it does not work at all with llvm 19 and 20.

Signed-off-by: Jan-Espen Oversand <sigsegv@radiotube.org>