net-vpn/vtun-ng: new package, add 3.0.16
This is a fork (by me) of net-vpn/vtun. It is a rewrite in rust
currently aiming to be a drop in replacement for net-vpn/vtun. Vtun
has not been maintained upstream since 2016, and it is being dropped
from other distros. I still think it is a useful tool so I finally
took the effort of forking and rewriting.

Signed-off-by: Jan-Espen Oversand <sigsegv@radiotube.org>