app-backup/b2: match upstream dependency versions
Updates both RDEPEND and DEPEND to match the upstream dependency ranges;
notably, b2 doesn't build with docutils-0.22 because it removes the
deprecated `docutils.utils.error_reporting` module (see [1]), and it was
pinned out in version 4.4.1. (There's an upstream issue tracking
this[2], but it has yet to be addressed.)

The versions specified for the `test` USE flag aren't an exact match for
upstream (`>=` instead of `~`) because most of the exact versions
specified no longer have ebuilds matching them.

[1] https://github.com/Backblaze/B2_Command_Line_Tool/issues/1101
[1] https://github.com/Backblaze/B2_Command_Line_Tool/issues/1103

Signed-off-by: Itai Ferber <itai@itaiferber.net>