www-apps/soupault: update SRC_URI, fix S variable
There's no such thing as codeberg.com. The ebuild was initially written
using github.com in the SRC_URI, but I prefer codeberg.org if
possible. The top-level folder in codeberg tarballs is the PN and not P,
so we need to fix the S variable to make the build work.

Signed-off-by: Thomas Bracht Laumann Jespersen <t@laumann.xyz>