media-sound/loudgain: add 0.6.8_p20240128-r3
Use v0.6.8 release archive and apply manually extracted patches instead
of downloading a commit archive from GitHub. This makes it possible to
share the source archive between 0.6.8-rN and 0.6.8_p20240128-rN.
I'm a not big fan of having lots of patches in files/ directory instead
of using an upstream commit archive, but loudgain repository (and a
commit archives as well) have prebuilt binary inside it, so it would be
beneficial to use fixed version of release archive. I feel that the
amount of patches are still acceptable.

Bump minimum required CMake version. See bug #951350.

Add a new rgbpm2 script, which is multi-processing version of rgbpm.
This was added to the upstream master branch after 0.6.8, but not yet
released.

I believe this ebuild can replace loudgain-0.6.8_p20240128-r2.ebuild
(and also loudgain-0.6.8-r2.ebuild), but considering the amount of
ebuild changes, I will keep those old ebuilds for a while.

Signed-off-by: YOSHIOKA Takuma <lo48576@hard-wi.red>